feat(trace): UI lazy loading of spans
move components into a separate component
