+++ app/src/components/trace/index.tsx
@@ -1,0 +1,1 @@
+export * from "./SpanKindIcon";
